auth-service/requirements.txt
django==5.0
djangorestframework==3.15
djangorestframework-simplejwt==5.3
psycopg2-binary
redis
django-cors-headers
gunicorn
python-dotenv
